feat: ChatInfo 정보 UserFamily로 병합
